requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
openai>=1.82.0
python-dotenv
PyYAML
httpx
requests
